+++ R/deps.R
@@ -541,6 +541,8 @@ parse_one_extra <- function(x, ...) {
     stop("Malformed remote specification '", x, "'", call. = FALSE)
   }
 
+  type = sub("^[.a-zA-Z0-9]+=", "", type)
+
   if (grepl("@", type)) {
     # Custom host
     tah <- strsplit(type, "@", fixed = TRUE)[[1]]